+++ src/paging.rs
@@ -13,6 +13,8 @@ use core::fmt::{self, Debug, Display, Formatter};
 use core::marker::PhantomData;
 use core::ops::{Add, Range, Sub};
 use core::ptr::NonNull;
+#[cfg(feature = "zerocopy")]
+use zerocopy::{AsBytes, FromBytes, FromZeroes};
 
 const PAGE_SHIFT: usize = 12;
 
@@ -62,7 +64,9 @@ impl TranslationRegime {
 }
 
 /// An aarch64 virtual address, the input type of a stage 1 page table.
+#[cfg_attr(feature = "zerocopy", derive(AsBytes, FromBytes, FromZeroes))]
 #[derive(Copy, Clone, Eq, Ord, PartialEq, PartialOrd)]
+#[repr(transparent)]
 pub struct VirtualAddress(pub usize);
 
 impl Display for VirtualAddress {
@@ -107,7 +111,9 @@ pub struct MemoryRegion(Range<VirtualAddress>);
 
 /// An aarch64 physical address or intermediate physical address, the output type of a stage 1 page
 /// table.
+#[cfg_attr(feature = "zerocopy", derive(AsBytes, FromBytes, FromZeroes))]
 #[derive(Copy, Clone, Eq, Ord, PartialEq, PartialOrd)]
+#[repr(transparent)]
 pub struct PhysicalAddress(pub usize);
 
 impl Display for PhysicalAddress {
@@ -828,6 +834,7 @@ impl<T: Translation> PageTableWithLevel<T> {
 
 /// A single level of a page table.
 #[repr(C, align(4096))]
+#[cfg_attr(feature = "zerocopy", derive(AsBytes, FromZeroes))]
 pub struct PageTable {
     entries: [Descriptor; 1 << BITS_PER_LEVEL],
 }
@@ -849,6 +856,7 @@ impl PageTable {
 ///   - A page mapping, if it is in the lowest level page table.
 ///   - A block mapping, if it is not in the lowest level page table.
 ///   - A pointer to a lower level pagetable, if it is not in the lowest level page table.
+#[cfg_attr(feature = "zerocopy", derive(AsBytes, FromZeroes))]
 #[derive(Clone, Copy, PartialEq, Eq)]
 #[repr(C)]
 pub struct Descriptor(usize);